Refactor into a package
