+++ setup.cfg
@@ -1,0 +1,15 @@
+[flake8]
+max-line-length = 120
+max-complexity = 10
+
+[isort]
+atomic = true
+combine_as_imports = false
+indent = 4
+length_sort = false
+line_length = 120
+multi_line_output = 5
+not_skip = __init__.py
+order_by_type = false
+wrap_length = 120
+known_third_party=six
