+++ tf_mnist/model.py
@@ -1,0 +1,104 @@
+import tensorflow as tf
+
+IMAGE_SIZE = 28
+IMAGE_SIZE_SQUARED = IMAGE_SIZE * IMAGE_SIZE
+
+
+# We can't initialize these variables to 0 - the network will get stuck.
+def weight_variable(shape):
+    """Create a weight variable with appropriate initialization."""
+    initial = tf.truncated_normal(shape, stddev=0.1)
+    return tf.Variable(initial)
+
+
+def bias_variable(shape):
+    """Create a bias variable with appropriate initialization."""
+    initial = tf.constant(0.1, shape=shape)
+    return tf.Variable(initial)
+
+
+def variable_summaries(var):
+    """Attach a lot of summaries to a Tensor (for TensorBoard visualization)."""
+    with tf.name_scope('summaries'):
+        mean = tf.reduce_mean(var)
+        tf.summary.scalar('mean', mean)
+        with tf.name_scope('stddev'):
+            stddev = tf.sqrt(tf.reduce_mean(tf.square(var - mean)))
+        tf.summary.scalar('stddev', stddev)
+        tf.summary.scalar('max', tf.reduce_max(var))
+        tf.summary.scalar('min', tf.reduce_min(var))
+        tf.summary.histogram('histogram', var)
+
+
+def make_model(learning_rate):
+    # Create a multilayer model.
+
+    # Input placeholders
+    with tf.name_scope('input'):
+        x = tf.placeholder(tf.float32, [None, IMAGE_SIZE_SQUARED], name='x-input')
+        y_ = tf.placeholder(tf.int64, [None], name='y-input')
+
+    with tf.name_scope('input_reshape'):
+        image_shaped_input = tf.reshape(x, [-1, IMAGE_SIZE, IMAGE_SIZE, 1])
+        tf.summary.image('input', image_shaped_input, 10)
+
+    all_weights = []
+    all_biases = []
+
+    def nn_layer(input_tensor, input_dim, output_dim, layer_name, act=tf.nn.relu):
+        """Reusable code for making a simple neural net layer.
+
+        It does a matrix multiply, bias add, and then uses relu to nonlinearize.
+        It also sets up name scoping so that the resultant graph is easy to read,
+        and adds a number of summary ops.
+        """
+        # Adding a name scope ensures logical grouping of the layers in the graph.
+        with tf.name_scope(layer_name):
+            # This Variable will hold the state of the weights for the layer
+            with tf.name_scope('weights'):
+                weights = weight_variable([input_dim, output_dim])
+                variable_summaries(weights)
+                all_weights.append(weights)
+            with tf.name_scope('biases'):
+                biases = bias_variable([output_dim])
+                variable_summaries(biases)
+                all_biases.append(biases)
+            with tf.name_scope('Wx_plus_b'):
+                preactivate = tf.matmul(input_tensor, weights) + biases
+                tf.summary.histogram('pre_activations', preactivate)
+            activations = act(preactivate, name='activation')
+            tf.summary.histogram('activations', activations)
+            return activations
+
+    hidden1 = nn_layer(x, IMAGE_SIZE_SQUARED, 500, 'layer1')
+
+    with tf.name_scope('dropout'):
+        keep_prob = tf.placeholder(tf.float32)
+        tf.summary.scalar('dropout_keep_probability', keep_prob)
+        dropped = tf.nn.dropout(hidden1, keep_prob)
+
+    y = nn_layer(dropped, 500, 10, 'layer2', act=tf.identity)
+
+    with tf.name_scope('cross_entropy'):
+        with tf.name_scope('total'):
+            cross_entropy = tf.losses.sparse_softmax_cross_entropy(labels=y_, logits=y)
+
+    final_tensor = tf.nn.softmax(logits=y, name='output')
+
+    tf.summary.scalar('cross_entropy', cross_entropy)
+
+    with tf.name_scope('train'):
+        train_step = tf.train \
+            .AdamOptimizer(learning_rate) \
+            .minimize(cross_entropy)
+
+    with tf.name_scope('accuracy'):
+        with tf.name_scope('correct_prediction'):
+            correct_prediction = tf.equal(tf.argmax(y, 1), y_)
+        with tf.name_scope('accuracy'):
+            accuracy = tf.reduce_mean(tf.cast(correct_prediction, tf.float32))
+
+    tf.summary.scalar('accuracy', accuracy)
+    merged = tf.summary.merge_all()
+
+    return locals()  # Hacky -- would be better to return just the named objects
