+++ tf_mnist/utils.py
@@ -1,0 +1,6 @@
+import os
+
+
+def get_first_file(path):
+    filename = os.listdir(path)[0]
+    return os.path.join(path, filename)
